backend/tests/test_retrieve.py
"""
Unit and integration tests for the RAG retrieval and validation system.
"""
import pytest
from unittest.mock import Mock, patch
import sys
import os

# Add the backend directory to the path so we can import retrieve
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))

# TODO: Tests will be implemented as the retrieve.py module is developed
# This file will contain:
# - Unit tests for Cohere integration
# - Unit tests for Qdrant client integration
# - Unit tests for data model validation
# - Integration tests for retrieval functionality
# - Integration tests for validation functionality